requirements.txt
opencv-python==4.7.0.72
numpy==1.24.3
pandas==2.0.1
easyocr==1.7.0
sqlalchemy==2.0.15
PyYAML==6.0
ultralytics==8.0.120
torch==2.0.1
torchvision==0.15.2